requirements.txt
asgiref==3.8.1
Django==5.1.2
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.10
sqlparse==0.5.1
